Refactored code using code inspection.

+++ src/test/java/org/beamproject/server/util/SessionStorageTest.java
@@ -21,9 +21,12 @@ package org.beamproject.server.util;
 import org.beamproject.common.Session;
 import org.beamproject.common.User;
 import org.beamproject.common.util.ComparableBytes;
-import org.junit.Test;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
 import org.junit.Before;
+import org.junit.Test;
 
 public class SessionStorageTest {
 
@@ -73,7 +76,7 @@ public class SessionStorageTest {
         assertFalse(storage.isSessionExisting(KEY));
 
         for (byte b : session.getKey()) {
-            assertEquals((byte) 0, b);
+            assertEquals(0, b);
         }
     }
